# back-end/recharge-api/src/main/resources/openapi/recharge-api.yaml
openapi: 3.0.3
info:
  title: Mobile Phone Recharge API
  description: API REST de recarga de celular com autenticação JWT
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local

tags:
  - name: Users
    description: Registro e gerenciamento de conta do usuário
  - name: Auth
    description: Autenticação (login, refresh, logout)
  - name: Phones
    description: Gerenciamento de telefones do usuário
  - name: Recharges
    description: Operações de recarga de celular

paths:
  /api/users:
    post:
      tags:
        - Users
      summary: Registrar novo usuário
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: E-mail já cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/me:
    get:
      tags:
        - Users
      summary: Dados do usuário logado com telefones vinculados
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Não autenticado
    put:
      tags:
        - Users
      summary: Atualizar dados do usuário
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
        '409':
          description: E-mail já em uso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Users
      summary: Excluir conta do usuário
      operationId: deleteCurrentUser
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Conta excluída com sucesso
        '401':
          description: Não autenticado
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login — retorna JWT
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Renovar token JWT
      operationId: refreshToken
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Token inválido ou expirado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout — limpa cookies JWT
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout bem-sucedido
        '401':
          description: Não autenticado

  /api/phones:
    post:
      tags:
        - Phones
      summary: Registrar um novo telefone
      operationId: registerPhone
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPhoneRequest'
      responses:
        '201':
          description: Telefone registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
        '409':
          description: Telefone já cadastrado para este usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Phones
      summary: Listar telefones do usuário
      operationId: listPhones
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de telefones do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PhoneResponse'
        '401':
          description: Não autenticado

  /api/phones/{id}:
    delete:
      tags:
        - Phones
      summary: Remover um telefone
      operationId: deletePhone
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Telefone removido com sucesso
        '401':
          description: Não autenticado
        '404':
          description: Telefone não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/recharges:
    post:
      tags:
        - Recharges
      summary: Criar recarga
      operationId: createRecharge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRechargeRequest'
      responses:
        '202':
          description: Recarga aceita para processamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RechargeResponse'
        '400':
          description: Dados inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Não autenticado
    get:
      tags:
        - Recharges
      summary: Histórico de recargas do usuário (paginado)
      operationId: getRechargeHistory
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de recargas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RechargePageResponse'
        '401':
          description: Não autenticado

  /api/recharges/{id}:
    get:
      tags:
        - Recharges
      summary: Detalhe de uma recarga
      operationId: getRechargeById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes da recarga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RechargeResponse'
        '401':
          description: Não autenticado
        '404':
          description: Recarga não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: João Silva
        email:
          type: string
          format: email
          example: joao@email.com
        password:
          type: string
          minLength: 6
          maxLength: 100
          format: password
          example: senha123

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: joao@email.com
        password:
          type: string
          format: password
          example: senha123

    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string

    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phones:
          type: array
          items:
            $ref: '#/components/schemas/PhoneResponse'

    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
          maxLength: 100
          format: password

    TokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresIn
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          description: Tempo de expiração em segundos

    CreateRechargeRequest:
      type: object
      required:
        - phoneNumber
        - amount
      properties:
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{10,14}$'
          example: "+5511999999999"
        amount:
          type: number
          format: double
          minimum: 10.0
          maximum: 200.0
          example: 30.00

    RechargeResponse:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        amount:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/RechargeStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RechargeStatus:
      type: string
      enum:
        - PENDING
        - PROCESSING
        - COMPLETED
        - FAILED

    RechargePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RechargeResponse'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time

    RegisterPhoneRequest:
      type: object
      required:
        - phoneNumber
        - label
      properties:
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{10,14}$'
          example: "+5511999999999"
        label:
          type: string
          minLength: 1
          maxLength: 50
          example: "Pessoal"

    PhoneResponse:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        label:
          type: string
        amount:
          type: number
          format: double
          example: 20.00
        createdAt:
          type: string
          format: date-time